various changes for new sql based files database

+++ src/operations/merge.py
@@ -19,7 +19,6 @@ import os
 import time
 import shutil
 import cPickle as pickle
-import xml.etree.cElementTree as iks
 
 import lpms
 
@@ -47,7 +46,9 @@ class Merge(internals.InternalFuncs):
         self.env = environment
         self.instdb = dbapi.InstallDB()
         self.conf = conf.LPMSConfig()
+        self.previous_files = []
         self.merge_conf_data = []
+        self.filesdb = dbapi.FilesDB()
         self.file_relationsdb = dbapi.FileRelationsDB()
         self.reverse_dependsdb = dbapi.ReverseDependsDB()
 
@@ -56,6 +57,15 @@ class Merge(internals.InternalFuncs):
             self.env.real_root = cst.root
         
         self.merge_conf_file = os.path.join(self.env.real_root, cst.merge_conf_file)
+
+        self.get_previous_files()
+
+    def get_previous_files(self):
+        version_data = self.instdb.get_version(self.env.name, pkg_category = self.env.category)
+        for slot in version_data:
+            if slot == self.env.slot:
+                self.previous_files = self.filesdb.get_paths_by_package(self.env.name, \
+                        repo=self.env.repo, category=self.env.category, version=version_data[slot][0])
 
     def load_merge_conf_file(self):
         if os.path.isfile(self.merge_conf_file):
@@ -80,11 +90,7 @@ class Merge(internals.InternalFuncs):
         if isinstance(status, bool) or not status or \
             not self.env.slot in status[-1].keys():
                 return True
-
-        #map(lambda x: slf.versions.extend(x), status[-1].values())
-        #if not status or not self.env.version in self.versions:
-        #    return True
-        
+      
         return False
 
     def is_reinstall(self):
@@ -120,26 +126,17 @@ class Merge(internals.InternalFuncs):
             }
 
 
-        conflict_check = False
-        previous_vers = self.instdb.get_version(self.env.name, pkg_category = self.env.category)
-        if previous_vers and self.env.slot in previous_vers:
-            fdb = filesdb.FilesDB(self.env.category, self.env.name, \
-                    previous_vers[self.env.slot][0])
-            fdb.import_xml()
-            conflict_check = True
+        conflict_check = True
+        if not self.previous_files:
+            conflict_check = False
+        self.filesdb.delete_item_by_pkgdata(self.env.category, self.env.name, \
+                self.env.version, commit=True)
 
-        out.notify("merging the package to %s and creating %s.xml" % (self.env.real_root, self.env.fullname))
+        out.notify("merging the package to %s and creating database entries" % self.env.real_root)
         
-        # create an xml object
-        xml_root = iks.Element("content")
-
         # find content of the package
         for root_path, dirs, files in os.walk(self.env.install_dir, followlinks=True):
             root_path = root_path.split(self.env.install_dir)[1]
-
-            # write directories
-            root = iks.SubElement(xml_root, "node")
-            root.set("path", root_path)
 
             # create directories
             for d in dirs:
@@ -167,11 +164,37 @@ class Merge(internals.InternalFuncs):
                     shelltools.set_id(target, perms["uid"], perms["gid"])
                     shelltools.set_mod(target, perms["mod"])
 
-                dir_tag = iks.SubElement(root, "dir")
-                for key in perms.keys():
-                    dir_tag.set(key, str(perms[key]))
-                dir_tag.text = d
-
+                    self.filesdb.add_path(
+                            (self.env.repo, 
+                                self.env.category, 
+                                self.env.name,
+                                self.env.version, 
+                                target, 
+                                "dir", 
+                                None, 
+                                perms['gid'],
+                                perms['mod'], 
+                                perms['uid'], 
+                                None, 
+                                None
+                            )
+                    )
+                else:
+                    self.filesdb.add_path(
+                            (self.env.repo, 
+                                self.env.category, 
+                                self.env.name,
+                                self.env.version, 
+                                target, 
+                                "link", 
+                                None, 
+                                None,
+                                None, 
+                                None, 
+                                None, 
+                                os.path.realpath(path)
+                            )
+                    )
 
             # write regular files
             reserve_files = []
@@ -188,7 +211,6 @@ class Merge(internals.InternalFuncs):
                         if not rf.startswith("#"):
                             reserve_files.append(rf.strip())
 
-
             for f in files:
                 source = os.path.join(self.env.install_dir, root_path[1:], f)
                 target = os.path.join(self.env.real_root, root_path[1:], f)
@@ -202,13 +224,13 @@ class Merge(internals.InternalFuncs):
                         self.file_relationsdb.add_file((self.env.repo, self.env.category, self.env.name, \
                                 self.env.version, target, file_relations.get_depends(source)))
 
-                if conflict_check and os.path.isfile(target) and not fdb.has_path(target):
-                    # FIXME: Use an exception to exit
-                    if self.conf.conflict_protect and not lpms.getopt("--ignore-conflicts"):
-                        out.error("file conflict: %s" % target)
-                        lpms.terminate()
-                    else:
-                        out.warn("file conflict: %s" % target)
+                #if conflict_check and os.path.isfile(target) and not fdb.has_path(target):
+                #    # FIXME: Use an exception to exit
+                #    if self.conf.conflict_protect and not lpms.getopt("--ignore-conflicts"):
+                #        out.error("file conflict: %s" % target)
+                #        lpms.terminate()
+                #    else:
+                #        out.warn("file conflict: %s" % target)
 
                 # strip binary files
                 if isstrip and utils.get_mimetype(source) in ('application/x-executable', 'application/x-archive', \
@@ -240,24 +262,43 @@ class Merge(internals.InternalFuncs):
                                 
                     if os.path.exists(target) and is_reserve():
                         #FIXME: Code duplication!!!
-                        # the file is reserved. Use existing version and create necessary XML stuff
+                        # the file is reserved.
                         if not os.path.islink(target):
                             shelltools.set_id(target, perms["uid"], perms["gid"])
                             shelltools.set_mod(target, perms["mod"])
-
-                        file_tag = iks.SubElement(root, "file")
-                        sha1sum = utils.sha1sum(target)
-
-                        for key in perms:
-                            file_tag.set(key, str(perms[key]))
-
-                        if sha1sum is not False:
-                            file_tag.set("sha1sum", sha1sum)
-                        size = utils.get_size(target)
-                        self.total += size
-                        file_tag.set("size", str(size))
-                        file_tag.text = f
-                    
+                            
+                            sha1sum = utils.sha1sum(target)
+                            self.filesdb.add_path(
+                                    (self.env.repo, 
+                                        self.env.category, 
+                                        self.env.name,
+                                        self.env.version, 
+                                        target,
+                                        "file", 
+                                        utils.get_size(source, dec=True), 
+                                        perms['gid'],
+                                        perms['mod'], 
+                                        perms['uid'], 
+                                        sha1sum, 
+                                        None
+                                    )
+                            )
+                        else:
+                            self.filesdb.add_path(
+                                    (self.env.repo, 
+                                        self.env.category, 
+                                        self.env.name,
+                                        self.env.version, 
+                                        target, 
+                                        "link", 
+                                        None, 
+                                        None,
+                                        None, 
+                                        None, 
+                                        None, 
+                                        os.path.realpath(path)
+                                    )
+                            )                   
                         # We don't need the following operations
                         continue
 
@@ -280,39 +321,43 @@ class Merge(internals.InternalFuncs):
                     shelltools.set_id(target, perms["uid"], perms["gid"])
                     shelltools.set_mod(target, perms["mod"])
 
-                file_tag = iks.SubElement(root, "file")
+                    self.filesdb.add_path(
+                            (self.env.repo, 
+                                self.env.category, 
+                                self.env.name,
+                                self.env.version, 
+                                target, 
+                                "file", 
+                                utils.get_size(source, dec=True), 
+                                perms['gid'],
+                                perms['mod'], 
+                                perms['uid'], 
+                                sha1sum, 
+                                None
+                            )
+                    )
+                else:
+                    self.filesdb.add_path(
+                            (self.env.repo, 
+                                self.env.category, 
+                                self.env.name,
+                                self.env.version, 
+                                target, 
+                                "link", 
+                                None, 
+                                None,
+                                None, 
+                                None, 
+                                None, 
+                                os.path.realpath(source)
+                            )
+                    )
 
-                for key in perms:
-                    file_tag.set(key, str(perms[key]))
-
-                if sha1sum is not False:
-                    file_tag.set("sha1sum", sha1sum)
-                size = utils.get_size(source)
-                self.total += size
-                file_tag.set("size", str(size))
-                file_tag.text = f
-
-        # set installed size
-        xml_root.set("size", str(self.total))
-
-        # fix indentation
-        utils.indent(xml_root)
-
-        # write the xml file to filesdb
-        self.filesdb_path = os.path.join(self.env.real_root, cst.db_path[1:], 
-                cst.filesdb, self.env.category, self.env.name)
-        shelltools.makedirs(self.filesdb_path)
-        self.myfile = self.filesdb_path+"/"+self.env.fullname+".xml.new"
-        if os.path.isfile(self.myfile):
-            shelltools.remove_file(self.myfile)
-        shelltools.echo(iks.tostring(xml_root, encoding='UTF-8'), self.myfile)
-        
         self.file_relationsdb.commit()
         lpms.logger.info("%s/%s merged to %s" % (self.env.category, self.env.fullname, \
                 self.env.real_root))
 
-        # it may be too big
-        del xml_root
+        self.filesdb.commit()
 
     def write_db(self):
         # write metadata
@@ -417,57 +462,35 @@ class Merge(internals.InternalFuncs):
         self.instdb.add_buildinfo(data)
 
     def clean_previous(self):
-        def obsolete_content():
-            version_data = self.instdb.get_version(self.env.name, pkg_category = self.env.category)
-            for slot in version_data:
-                if slot == self.env.slot:
-                    return self.comparison(self.env.version, version_data[slot][0])
-
         if not self.is_fresh():
             if self.is_different() or self.is_reinstall():
-                obsolete_dirs, obsolete_files = obsolete_content()
+                obsolete = self.comparison()
 
-                if len(obsolete_dirs) != 0 or len(obsolete_files) !=0:
-                    out.normal("cleaning obsolete content...")
-                
-                for _file in obsolete_files:
-                    target = os.path.join(self.env.real_root, _file[1:])
+                if not obsolete: return
+
+                out.normal("cleaning obsolete content")
+                dirs = []
+                for item in obsolete:
+                    target = os.path.join(self.env.real_root, item[1:])
                     if os.path.islink(target):
                         os.unlink(target)
-                    else:
+                    elif os.path.isfile(target):
                         shelltools.remove_file(target)
+                    else:
+                        dirs.append(target)
 
-                dirs = obsolete_dirs
                 dirs.reverse()
-                for _dir in dirs:
-                    target = os.path.join(self.env.real_root, _dir[1:])
-                    if len(os.listdir(target)) == 0:
-                        shelltools.remove_dir(target)
-                        
-        shelltools.rename(self.myfile, self.myfile.split(".new")[0])
+                for item in dirs:
+                    if not os.listdir(item): shelltools.remove_dir(item)
 
-    def comparison(self, new_ver, old_ver):
-        obsolete_dirs = []; obsolete_files = []
-        new = filesdb.FilesDB(self.env.category, 
-                self.env.name, new_ver, self.env.real_root, suffix=".xml.new")
-        new.import_xml()
-
-        old = filesdb.FilesDB(self.env.category,
-                self.env.name, old_ver, self.env.real_root)
-        old.import_xml()
-
-        # for directories
-        for _dir in old.content['dirs']:
-            if not _dir in new.content['dirs']:
-                obsolete_dirs.append(_dir)
-
-        # for regular files
-        for _file in old.content['file']:
-            if not _file in new.content['file']:
-                obsolete_files.append(_file)
-
-        shelltools.remove_file(os.path.join(self.filesdb_path, self.env.name)+"-"+old_ver+".xml")
-        return obsolete_dirs, obsolete_files
+    def comparison(self):
+        current_files = self.filesdb.get_paths_by_package(self.env.name, \
+                        repo=self.env.repo, category=self.env.category, version=self.env.version)
+        obsolete = []
+        for item in self.previous_files:
+            if not item in current_files:
+                obsolete.append(item)
+        return obsolete
 
 def main(environment):
     opr = Merge(environment)
